+++ multi/integration/loveManager/extensions.lua
@@ -100,6 +100,32 @@ function multi:newSystemThreadedJobQueue(n)
 		self.queue:push{name,self.id,...}
 		return self.id
 	end
+	local nFunc = 0
+    function c:newFunction(name,func,holup) -- This registers with the queue
+        if type(name)=="function" then
+            holup = func
+            func = name
+            name = "JQ_Function_"..nFunc
+        end
+        nFunc = nFunc + 1
+        c:registerFunction(name,func)
+        return thread:newFunction(function(...)
+            local id = c:pushJob(name,...)
+            local link
+            local rets
+            link = c.OnJobCompleted(function(jid,...)
+                if id==jid then
+                    rets = {...}
+                    link:Remove()
+                end
+            end)
+            return thread.hold(function()
+                if rets then
+                    return unpack(rets)
+                end
+            end)
+        end,holup),name
+    end
 	multi:newThread("jobManager",function()
 		while true do
 			thread.yield()